requirements.txt
openai>=1.40
langchain>=0.2
langchain-community>=0.2
pydantic>=2.7
tenacity>=8.2
streamlit>=1.35
pandas>=2.2
plotly>=5.22
faiss-cpu>=1.8
tiktoken>=0.7
python-dotenv>=1.0
sqlmodel>=0.0.21
ruff>=0.4.4
mypy>=1.10
sqlalchemy>=2.0.0
requests>=2.32
reportlab>=4.2
